LPS-54752 Move internal methods to a helper class so that in upcoming commits we can use it from other classes
